+++ types/common/chain.go
@@ -1,6 +1,7 @@
 package common
 
 import (
+	"fmt"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -11,7 +12,13 @@ func (chain *Chain) ID() *big.Int {
 	if chain.Id == "" {
 		return big.NewInt(0)
 	}
-	return hexutil.MustDecodeBig(chain.Id)
+
+	chainID, ok := big.NewInt(0).SetString(chain.Id, 10)
+	if !ok {
+		panic(fmt.Sprintf("invalid decimal chain ID %q", chain.Id))
+	}
+
+	return chainID
 }
 
 // SetID set chain ID